apps/ml/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
greenlet>=3.1.1
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
aiosqlite>=0.20.0
redis>=5.2.0
arq>=0.26.1
httpx>=0.28.0
boto3>=1.35.0
tenacity>=9.0.0
orjson>=3.10.0
sse-starlette>=2.1.3
prometheus-client>=0.21.0
PyYAML>=6.0.2
python-multipart>=0.0.19
python-pptx>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
Pillow>=11.0.0
matplotlib>=3.9.0
edge-tts>=6.1.18
faster-whisper>=1.1.0
python-docx>=1.1.2
reportlab>=4.2.5
pytest>=8.3.3
pytest-asyncio>=0.24.0
